+++ client/src/utils/validation/userValidation.ts
@@ -10,7 +10,7 @@ export function validateEmail(value: string): string | null {
   }
 
   const regex: RegExp =
-    /^(?=.{6,254}$)[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]{0,64}@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?\.[a-zA-Z]{2,}?(?:\.[a-zA-Z]{2,})*$/;
+    /^(?=.{1,254}$)[a-zA-Z0-9!#$%&'*+/=?^_`{|}~.-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
   if (!regex.test(value)) {
     return 'Invalid email format.';
   }