+++ src/util/sqlUtils/generatePlaceHolders.ts
@@ -1,14 +1,3 @@
 export function generatePlaceHolders(numberOfPlaceHolders: number): string {
-  let placeHolderString: string = '';
-
-  for (let i = 0; i < numberOfPlaceHolders; i++) {
-    if (i + 1 === numberOfPlaceHolders) {
-      placeHolderString += '?';
-      continue;
-    }
-
-    placeHolderString += '?, ';
-  }
-
-  return placeHolderString;
+  return Array(numberOfPlaceHolders).fill('?').join(', ');
 }